del old gui_class rename new one to it
