+++ GUI_class.py
@@ -5,6 +5,8 @@ Jan 2019
 #======================
 # imports
 #======================
+from BankReconciliation import BankReconciliation
+import ast
 import pandas as pd
 import numpy as np
 import warnings
@@ -12,25 +14,41 @@ warnings.filterwarnings('ignore')
 import tkinter as tk
 from tkinter import ttk, scrolledtext, Menu, \
                     messagebox as msg, Spinbox, \
-                    filedialog
+                    filedialog, MULTIPLE, EXTENDED
 from time import  sleep         # careful - this can freeze the GUI
 global sol,f1Var,filePathBank,\
         filePathLedger,filePathBank, \
         intRad, intChk
 filePathBank = ""
 filePathLedger = ""
-# bank = pd.read_csv("~/Documents/GitHub/senior-bank-reconcile/example-data/bank-v2.csv")
-result_bank = pd.read_csv("~/Documents/GitHub/senior-bank-reconcile/example-data/bank-test")
-# ledger = pd.read_csv("~/Documents/GitHub/senior-bank-reconcile/example-data/ledger-v2.csv")
-result_ledger = pd.read_csv("~/Documents/GitHub/senior-bank-reconcile/example-data/ledger-test")
+# bankDF = pd.read_csv("~/Documents/GitHub/senior-bank-reconcile/example-data/bank-v2.csv")
+# result_bank = pd.read_csv("~/Documents/GitHub/senior-bank-reconcile/example-data/bank-test")
+# ledgerDF = pd.read_csv("~/Documents/GitHub/senior-bank-reconcile/example-data/ledger-v2.csv")
+# result_ledger = pd.read_csv("~/Documents/GitHub/senior-bank-reconcile/example-data/ledger-test")
 #=================================================================== 
-class GUI():
-    def __init__(self):         # Initializer method
-        # Create instance
+class GUI(BankReconciliation):
+    def __init__(self):
+        self.setupWindow()
+        self.initUploadScreen()
+        self.initUI()
+        bankDF = pd.read_csv(filePathBank)
+        ledgerDF = pd.read_csv(filePathLedger)
+        self.reconciled = BankReconciliation(bankDF,ledgerDF)
+        # ***** Static Column Name Problem
+        tempCol = self.reconciled.bankDF['associate']
+        tempCol.fillna(-1,inplace=True)
+        tempCol = tempCol.astype(np.int64)
+        print(tempCol)
+        sol = list(tempCol)
+
+        self.set_solution(sol)
+        self.setupWindow()
+        self.planning_data(self.reconciled.bankDF,self.reconciled.ledgerDF,sol)
+        self.initUI()
+
+    def setupWindow(self):
         self.win = tk.Tk()   
-        
-        # Add a title       
-        self.win.title("Python GUI")      
+        self.win.title("Python GUI")
 
     def set_solution(self, solution):
         global sol
@@ -41,10 +59,13 @@ class GUI():
     
     def clickUpload(self):
         self.__filePath = filedialog.askopenfilename()
-        self.enteredfilePath.configure(text= self.__filePath)
-        global filePathBank
-        filePathBank = self.__filePath
-        print("filePath = " + filePathBank)
+        if self.__filePath is None:
+            return
+        else:
+            self.enteredfilePath.configure(text= self.__filePath)
+            global filePathBank
+            filePathBank = self.__filePath
+            print("filePath = " + filePathBank)
 
     def clickUpload2(self):
         self.__filePath = filedialog.askopenfilename()
@@ -56,44 +77,124 @@ class GUI():
             filePathLedger = self.__filePath
             print("filePath2 = " + filePathLedger)
 
-    def NextClicked(self):
+    def nextClicked(self):
         self.win.destroy()
 
-    def selecting_button(self, solution):
-        for i in range(len(solution)):
-            if solution[i] is not None:
-                __temp = "variable{0}".format(solution[i])
-                self.d2[__temp+"_button"].select()
-                print(__temp+"_button selected")
+    def editTransaction(self):
+        t = tk.Toplevel(self.win)
+        t.wm_title("Window")
+        tabControl = ttk.Notebook(t)
+        # Tab Match
+        def saveSelection():
+            __str = lb1.curselection()
+            resultSelect.configure(text=str(__str))
+            indexBank = self.f1Var.get()
+            sol[indexBank] = list(__str)
 
-    def InteractedRadioButton(self):
-        '''GUI function used in planning_data function,
-        radiobutton so it can auto disabled/active checkbox on Ledger side
-        '''
-        for index,target in enumerate(sol):
-            # print("interacting...",end="")
-            __str2 = "variable{0}".format(target)
-            __value=self.f1Var.get()
-            global intRad
-            intRad = __value
-            # print(index, ":",target,end=",")
-            if __value == index:
-                try:
-                    self.d2[__str2+"_button"].configure(state="active")
-                except KeyError as identifier:
-                    continue
-            elif target is not None:
-                self.d2[__str2+"_button"].configure(state="disabled") 
-            else:
-                continue
-        print()
-    
-    def InteractedCheckbox(self):
-        global intChk
+            # Senting updated value to  MainPage
+            __str2 = "update{0}".format(indexBank)
+            __str3 = """"""
+            for value in sol[indexBank]:
+                row = self.reconciled.ledgerDF.iloc[value]
+                # Static Name Problem
+                __temp = str(str(row["Date"])+\
+                    ","+str (row["Item"])+\
+                    ","+str(row["Debit"])+\
+                    ","+str(row["Credit"])+\
+                    ","+str(row["Balance"]))
+                __str3 = __str3 + __temp + '\n'
+                # print(__str3)
+            
+            d2[__str2].configure(text=str(__str3))
+            print("Saved Selection", sol[indexBank], __str)
+            print(sol)
+        tab1 = ttk.Frame(tabControl)
+        tabControl.add(tab1,text="Match")
+        lb1 = tk.Listbox(tab1,selectmode=MULTIPLE,width=70)
+        for index, row in self.reconciled.ledgerDF.iterrows(): 
+            # Static Name Column Problem ****
+            lb1.insert(index,\
+                (row["Date"]\
+                ,row["Item"]\
+                ,row["Debit"]\
+                ,row["Credit"]\
+                ,row["Balance"]))
+        resultSelect = ttk.Label(t)
+        bt = ttk.Button(tab1, text="Save Selection", command=saveSelection)
 
-        pass
+        # Tab Create
+        def saveCreate():
+            __list = [what.get(), why.get(), who.get()]
+            indexBank = self.f1Var.get()
+            sol[indexBank] = __list
+            __str2 = "update{0}".format(indexBank)
+            d2[__str2].configure(text=str(__list))
+            print(__list)
+            print(sol)
+            pass
+        tab2 = ttk.Frame(tabControl)
+        tabControl.add(tab2,text="Create")
+
+        ## What
+        what = tk.StringVar()
+        tk.Label(tab2, text="What").pack()
+        ttk.Entry(tab2, textvariable=what).pack()
+        ## Why
+        why = tk.StringVar()
+        tk.Label(tab2, text="Why").pack()
+        ttk.Entry(tab2, textvariable=why).pack()
+        ## Who
+        who = tk.StringVar()
+        tk.Label(tab2, text="Who").pack()
+        ttk.Entry(tab2, textvariable=who).pack()
+
+        ttk.Button(tab2, text="Save Create", command=saveCreate).pack()
+
+        # Tab Transfer
+        def saveTransfer():
+            __list = [bankAccount.get(), reference.get()]
+            indexBank = self.f1Var.get()
+            sol[indexBank] = __list
+            __str2 = "update{0}".format(indexBank)
+            d2[__str2].configure(text=str(__list))
+            print(__list)
+            print(sol)            
+            pass
+        tab3 = ttk.Frame(tabControl)
+        tabControl.add(tab3,text="Transfer")
+        ## Bank Account
+        bankAccount = tk.StringVar()
+        tk.Label(tab3,text="Bank Acccount").pack()
+        ttk.Entry(tab3, textvariable=bankAccount).pack()
+        ## Reference
+        reference = tk.StringVar()
+        tk.Label(tab3,text="Reference").pack()
+        ttk.Entry(tab3,textvariable=reference).pack()
+        ttk.Button(tab3,text="Save Transfer",command=saveTransfer).pack()
+
+        # Tab Discuss
+        def saveDiscuss():
+            __list = [comment.get()]
+            indexBank = self.f1Var.get()
+            sol[indexBank] = __list
+            __str2 = "update{0}".format(indexBank)
+            d2[__str2].configure(text=str(__list))
+            print(__list)
+            print(sol)            
+            pass
+        comment = tk.StringVar()
+        tab4 = ttk.Frame(tabControl)
+        tabControl.add(tab4,text="Discuss") 
+        tk.Label(tab4, text="Comment").pack()
+        ttk.Entry(tab4,width=50,textvariable=comment).pack()
+        ttk.Button(tab4, text="Save Discuss", command=saveDiscuss).pack()      
+
+        lb1.pack()
+        resultSelect.pack()
+        bt.pack()
+        tabControl.pack(expan=1, fill="both")
+        self.win.mainloop()
         
-
     def planning_data(self,df,df2,solution):
         '''
         run this for GUI for after processing CSV file automated, 
@@ -104,28 +205,55 @@ class GUI():
         which represent (solution, index 0 contain 0) meaning ...
         ...(df row 0) are reconcile with (df2 row 0)
         '''
-        global d1;self.d1 = {}
-        global d2;self.d2 = {}
+        def saveToFile():
+            path = filedialog.asksaveasfilename()
+            if path is '':
+                pass
+            else:
+                _file = open(path,'w+')
+                _file.write(str(sol))
+                print(str(sol))
+                _file.close()
+
+        def loadFromFile():
+            path = filedialog.askopenfilename()
+            if path is '':
+                pass
+            else:
+                ("loading..." + str(path))
+                _file = open(path,'r')
+                txt = _file.read()
+                _file.close()
+                sol = ast.literal_eval(txt)
+                print("set sol = " + str(sol))
+
+            
+        global d1;d1 = {}
+        global d2;d2 = {}
         global sol; self.solution = solution
-        self.frame2 = ttk.LabelFrame(self.win, text="Layout 2")
-        self.frame2.grid(column=0, row=1) 
-        for index, row in df2.iterrows():
-            __str2 = "variable{0}".format(index)
-            self.d2[__str2]= tk.IntVar()
-            self.d2[__str2+"_button"] = tk.Checkbutton(self.frame2, variable=__str2, command=self.InteractedCheckbox)
-            self.d2[__str2+"_button"].grid(column=0, row=index, sticky=tk.W)
-            for i in range(len(df.columns)):
-                ttk.Label(self.frame2, text=str(row[i])).grid(column=i+1, row=index)
         self.frame1 = ttk.LabelFrame(self.win, text="Layout 1")
         self.frame1.grid(column=0, row=0)
         self.f1Var = tk.IntVar()
-        for index, row in df.iterrows(): 
+        for index, row in self.reconciled.bankDF.iterrows():
             __str = "variable{0}".format(index)
-            self.d1[__str+"_button"] = tk.Radiobutton(self.frame1, variable=self.f1Var, value = index ,command = self.InteractedRadioButton)
-            self.d1[__str+"_button"].grid(column=0, row=index, sticky=tk.W)
-            for i in range(len(df.columns)):
-                ttk.Label(self.frame1, text=str(row[i])).grid(column=i+1, row=index) 
-        self.selecting_button(solution)
+            d1[__str+"_button"] = tk.Radiobutton(self.frame1, variable=self.f1Var, value = index )
+            d1[__str+"_button"].grid(column=0, row=index, sticky=tk.W)
+            # Fixed Selection Name Problem
+            ttk.Label(self.frame1, text=str(row["Date"])).grid(column=1, row=index)
+            ttk.Label(self.frame1, text=str(row["Description"])).grid(column=2, row=index)
+            ttk.Label(self.frame1, text=str(row["Withdrawals"])).grid(column=3, row=index)
+            ttk.Label(self.frame1, text=str(row["Deposits"])).grid(column=4, row=index)
+            ttk.Label(self.frame1, text=str(row["Balance"])).grid(column=5, row=index)
+            ttk.Label(self.frame1, text=str(row["associate"])).grid(column=6, row=index)
+            __str2 = "update{0}".format(index)
+            d2[__str2] = ttk.Label(self.frame1)
+            d2[__str2].grid(column=7, row=index)
+        editButton = ttk.Button(self.win, text="Edit", command=self.editTransaction)
+        editButton.grid(column=0, row=1)
+        saveButton = ttk.Button(self.win, text="Save To File...", command=saveToFile)
+        saveButton.grid(column=0, row=2)
+        loadButton = ttk.Button(self.win, text="Load File...",command=loadFromFile)
+        loadButton.grid(column=0, row=3)
         print("Done! Planning Data!")
         pass
 
@@ -145,18 +273,10 @@ class GUI():
         self.enteredFilePath2 = tk.Label(self.frame2, width=50)
         self.enteredFilePath2.grid(column=0, row=0)
 
-        self.nextButton = ttk.Button(self.frame2, text="Next...", command=self.NextClicked)
+        self.nextButton = ttk.Button(self.frame2, text="Next...", command=self.nextClicked)
         self.nextButton.grid(column=0,row = 1)
-
         pass
 
-
-sol = [0,6,1,2,None,4,None,None,None,7]
-ui = GUI()
-
-# ui.initUploadScreen()
-# ui.initUI()
-
-ui.set_solution(sol)
-ui.planning_data(result_bank,result_ledger,sol)
-ui.initUI()
+# ui = GUI()
+if __name__ == "__main__":
+    ui = GUI()
